build: follow fastcdc 3 → 5 (FastCDC::new takes usize sizes)
fastcdc 5 changed `v2020::FastCDC::new` to take `usize` for the
min/avg/max block sizes. `ChunkParams` keeps its `u32` fields; the one
call site in `FilesystemStore::chunk` widens them with `as usize`,
which is lossless on every supported target.

Behaviour is unchanged: fastcdc 5 requires even sizes (debug_assert),
and every `ChunkParams` in the tree is an even power of two, for which
cut points are identical to 3.x. Content IDs are computed here with
blake3 (`Cid::of`), not fastcdc's fingerprint, so they are unaffected.

+++ services/fq-store/src/fs.rs
@@ -272,13 +272,19 @@ impl BlockStore for FilesystemStore {
         if content.is_empty() {
             return Vec::new();
         }
-        fastcdc::v2020::FastCDC::new(content, self.params.min, self.params.avg, self.params.max)
-            .map(|c| Chunk {
-                hash: Cid::of(&content[c.offset..c.offset + c.length]),
-                offset: c.offset,
-                len: c.length,
-            })
-            .collect()
+        // fastcdc 5 takes `usize` sizes; `ChunkParams` stays `u32` (lossless widen).
+        fastcdc::v2020::FastCDC::new(
+            content,
+            self.params.min as usize,
+            self.params.avg as usize,
+            self.params.max as usize,
+        )
+        .map(|c| Chunk {
+            hash: Cid::of(&content[c.offset..c.offset + c.length]),
+            offset: c.offset,
+            len: c.length,
+        })
+        .collect()
     }
 
     async fn write_block(&self, block: &Cid, generation: u32, bytes: &[u8]) -> Result<()> {
